chore(chunkserver): Move MasterConn to own files
This change just moves the MasterConn class to own files, to reduce the
complexity in the masterconn module for implementing later multiple
connections to MDSs. The change also aims for better separation of
concerns and modularity.
