+++ src/chunkserver/masterconn.cc
@@ -34,680 +34,25 @@
 #include <cstdlib>
 #include <cstring>
 #include <ctime>
-#include <list>
 
-#include "chunkserver-common/hdd_utils.h"
 #include "chunkserver/bgjobs.h"
 #include "chunkserver/hddspacemgr.h"
-#include "chunkserver/network_main_thread.h"
+#include "chunkserver/master_connection.h"
 #include "common/event_loop.h"
-#include "common/loop_watchdog.h"
 #include "common/massert.h"
 #include "common/network_address.h"
-#include "common/output_packet.h"
 #include "common/random.h"
 #include "common/sockets.h"
-#include "common/time_utils.h"
 #include "config/cfg.h"
 #include "devtools/request_log.h"
 #include "protocol/SFSCommunication.h"
 #include "protocol/cstoma.h"
-#include "protocol/input_packet.h"
-#include "protocol/matocs.h"
-#include "protocol/packet.h"
 #include "slogger/slogger.h"
-
-static constexpr uint32_t kMaxPacketSize = 10000;
-static constexpr uint32_t kMaxBackgroundJobsCount = 1000;
-
-// Common variables from config
-static std::string gBindHostStr;
-static bool gEnableLoadFactor;
-static std::string gLabel;
-static uint32_t gTimeout_ms;
-
-enum class ConnectionMode : std::uint8_t {
-	FREE,        /// There is no socket for the connection yet.
-	CONNECTING,  /// Connection is being established.
-	CONNECTED,   /// Connection is active.
-	KILL         /// Connection has been dropped, a reconnection will be attempted.
-};
-
-/// @brief Class representing a connection to a Metadata Server (MDS).
-///
-/// Currently, only one active MDS (known as Master) is supported.
-class MasterConn {
-public:
-	explicit MasterConn(const std::string &masterHostStr, const std::string &masterPortStr,
-	                    const std::shared_ptr<JobPool> &jobPool)
-	    : masterHostStr_(masterHostStr), masterPortStr_(masterPortStr), jobPool_(jobPool) {}
-
-	// Disable unneeded copying and moving of the connection objects.
-	MasterConn(const MasterConn &) = delete;
-	MasterConn(MasterConn &&) = delete;
-	MasterConn &operator=(const MasterConn &) = delete;
-	MasterConn &operator=(const MasterConn &&) = delete;
-
-	~MasterConn() {
-		if (socketFD_ >= 0) { tcpclose(socketFD_); }
-	}
-
-	// Packet handling
-
-	static void deletePacket(void *packet) {
-		auto *outputPacket = static_cast<OutputPacket *>(packet);
-		delete outputPacket;
-	}
-
-	void attachPacket(void *packet) {
-		auto *outputPacket = static_cast<OutputPacket *>(packet);
-		outputPackets_.emplace_back(std::move(*outputPacket));
-		delete outputPacket;
-	}
-
-	void createAttachedPacket(MessageBuffer serializedPacket) {
-		outputPackets_.emplace_back(std::move(serializedPacket));
-	}
-
-	template <class... Data>
-	void createAttachedNoVersionPacket(PacketHeader::Type type, const Data &...data) {
-		std::vector<uint8_t> buffer;
-		serializeLegacyPacket(buffer, type, data...);
-		createAttachedPacket(std::move(buffer));
-	}
-
-	// Configuration
-
-	void reloadConfig() {
-		auto newMasterHostStr_ = cfg_getstring("MASTER_HOST", "sfsmaster");
-		auto newMasterPortStr_ = cfg_getstring("MASTER_PORT", "9420");
-
-		if (newMasterHostStr_ == masterHostStr_ && newMasterPortStr_ == masterPortStr_) {
-			return;  // no change
-		}
-
-		masterHostStr_ = newMasterHostStr_;
-		masterPortStr_ = newMasterPortStr_;
-
-		uint32_t mip{};
-		uint16_t mport{};
-
-		if (tcpresolve(masterHostStr_.c_str(), masterPortStr_.c_str(), &mip, &mport, 0) >= 0) {
-			if (isLoopbackAddress(mip)) {
-				safs::log_warn(
-				    "Chunkserver loopback IP addresses are experimental; consider a non-loopback IP address to chunkserver (via /etc/hosts or some other way)");
-			}
-
-			if (address().ip != mip || address().port != mport) {
-				setMasterAddress(mip, mport);
-				setMode(ConnectionMode::KILL);
-			}
-		} else {
-			safs::log_warn("master connection module: can't resolve master host/port ({}:{})",
-			               masterHostStr_, masterPortStr_);
-		}
-	}
-
-	// Connection management
-
-	void sendRegisterLabel() {
-		if (mode_ == ConnectionMode::CONNECTED) {
-			createAttachedPacket(cstoma::registerLabel::build(gLabel));
-		}
-	}
-
-	void sendConfig() {
-		if (mode_ == ConnectionMode::CONNECTED) {
-			createAttachedPacket(cstoma::registerConfig::build(cfg_yaml_string()));
-		}
-	}
-
-	void sendRegister() {
-		uint32_t myip;
-		uint16_t myport;
-		uint64_t usedspace, totalspace;
-		uint64_t tdusedspace, tdtotalspace;
-		uint32_t chunkcount, tdchunkcount;
-
-		myip = mainNetworkThreadGetListenIp();
-		myport = mainNetworkThreadGetListenPort();
-		createAttachedPacket(
-		    cstoma::registerHost::build(myip, myport, gTimeout_ms, SAUNAFS_VERSHEX));
-
-		hddForeachChunkInBulks([this](const std::vector<ChunkWithVersionAndType> &chunksBulk) {
-			createAttachedPacket(cstoma::registerChunks::build(chunksBulk));
-		});
-
-		hddGetTotalSpace(&usedspace, &totalspace, &chunkcount, &tdusedspace, &tdtotalspace,
-		                 &tdchunkcount);
-		auto registerSpace = cstoma::registerSpace::build(usedspace, totalspace, chunkcount,
-		                                                  tdusedspace, tdtotalspace, tdchunkcount);
-		createAttachedPacket(std::move(registerSpace));
-		sendRegisterLabel();
-		sendConfig();
-	}
-
-	int initConnect() {
-		if (!isMasterAddressValid_) {
-			uint32_t mip{};
-			uint32_t bip{};
-			uint16_t mport{};
-
-			if (tcpresolve(gBindHostStr.c_str(), nullptr, &bip, nullptr, 1) < 0) { bip = 0; }
-
-			bindHostAddress_.ip = bip;
-
-			if (tcpresolve(masterHostStr_.c_str(), masterPortStr_.c_str(), &mip, &mport, 0) >= 0) {
-				if (isLoopbackAddress(mip)) {
-					safs::log_warn(
-					    "Chunkserver loopback IP addresses are experimental; consider assigning an IP address to chunkserver (via /etc/hosts or some other way)");
-				}
-				address_.ip = mip;
-				address_.port = mport;
-				isMasterAddressValid_ = true;
-			} else {
-				safs::log_warn("master connection module: can't resolve master host/port ({}:{})",
-				               masterHostStr_, masterPortStr_);
-				return -1;
-			}
-		}
-
-		socketFD_ = tcpsocket();
-
-		if (socketFD_ < 0) {
-			safs::log_error_code(errno, "master connection module: create socket error");
-			return -1;
-		}
-
-		if (tcpnonblock(socketFD_) < 0) {
-			safs::log_error_code(errno, "master connection module: set nonblock error");
-			tcpclose(socketFD_);
-			socketFD_ = -1;
-			return -1;
-		}
-
-		if (bindHostAddress_.ip > 0) {
-			if (tcpnumbind(socketFD_, bindHostAddress_.ip, 0) < 0) {
-				safs::log_error_code(errno, "master connection module: can't bind socket to given ip");
-				tcpclose(socketFD_);
-				socketFD_ = -1;
-				return -1;
-			}
-		}
-
-		int status = tcpnumconnect(socketFD_, address_.ip, address_.port);
-
-		if (status < 0) {
-			safs::log_error_code(errno, "master connection module: connect failed");
-			tcpclose(socketFD_);
-			socketFD_ = -1;
-			isMasterAddressValid_ = false;
-			return -1;
-		}
-
-		if (status == 0) {
-			safs::log_info("connected to Master immediately");
-			onConnected();
-		} else {
-			mode_ = ConnectionMode::CONNECTING;
-			safs::log_info("connecting to Master");
-		}
-
-		return 0;
-	}
-
-	void connectTest() {
-		int status = tcpgetstatus(socketFD_);
-
-		if (status) {
-			safs::log_error_code(errno, "connection failed");
-			tcpclose(socketFD_);
-			socketFD_ = -1;
-			mode_ = ConnectionMode::FREE;
-			isMasterAddressValid_ = 0;
-		} else {
-			safs::log_info("connected to Master");
-			onConnected();
-		}
-	}
-
-	void onConnected() {
-		tcpnodelay(socketFD_);
-		mode_ = ConnectionMode::CONNECTED;
-		inputPacket_.reset();
-
-		sendRegister();
-		lastRead_.reset();
-		lastWrite_.reset();
-	}
-
-	// Polling
-
-	void providePollDescriptors(std::vector<pollfd> &pdesc) {
-		pDescPos_ = -1;
-
-		if (mode_ == ConnectionMode::FREE || socketFD_ < 0) { return; }
-
-		if (mode_ == ConnectionMode::CONNECTED) {
-			if (jobPool_->getJobCount() < (kMaxBackgroundJobsCount * 9) / 10) {
-				pdesc.emplace_back(socketFD_, POLLIN, 0);
-				pDescPos_ = static_cast<int32_t>(pdesc.size() - 1);
-			}
-		}
-
-		if (((mode_ == ConnectionMode::CONNECTED) && !outputPackets_.empty()) ||
-		    mode_ == ConnectionMode::CONNECTING) {
-			if (pDescPos_ >= 0) {
-				pdesc[pDescPos_].events |= POLLOUT;
-			} else {
-				pdesc.emplace_back(socketFD_, POLLOUT, 0);
-				pDescPos_ = static_cast<int32_t>(pdesc.size() - 1);
-			}
-		}
-	}
-
-	void handlePollErrors(const std::vector<pollfd> &pdesc) {
-		// Check if the socket has been closed or has an error.
-		if (pDescPos_ >= 0 && (pdesc[pDescPos_].revents & (POLLHUP | POLLERR))) {
-			if (mode_ == ConnectionMode::CONNECTING) {
-				connectTest();
-			} else {
-				mode_ = ConnectionMode::KILL;
-			}
-		}
-	}
-
-	void servePoll(const std::vector<pollfd> &pdesc) {
-		if (mode_ == ConnectionMode::CONNECTING) {
-			// Check if the connection has been established.
-			if (socketFD_ >= 0 && pDescPos_ >= 0 && (pdesc[pDescPos_].revents & POLLOUT)) {
-				connectTest();
-			}
-		} else {
-			if (pDescPos_ >= 0) {
-				// Check if there is data to read from this connection
-				if ((mode_ == ConnectionMode::CONNECTED) && (pdesc[pDescPos_].revents & POLLIN)) {
-					lastRead_.reset();
-					readFromSocket();
-				}
-
-				// Check if there is data to write to this connection
-				if ((mode_ == ConnectionMode::CONNECTED) && (pdesc[pDescPos_].revents & POLLOUT)) {
-					lastWrite_.reset();
-					writeToSocket();
-				}
-
-				// Check if the connection has not been used for a while and should be closed
-				if ((mode_ == ConnectionMode::CONNECTED) && lastRead_.elapsed_ms() > gTimeout_ms) {
-					mode_ = ConnectionMode::KILL;
-				}
-
-				// Keep the connection alive by sending a NOP packet
-				if ((mode_ == ConnectionMode::CONNECTED) &&
-				    lastWrite_.elapsed_ms() > (gTimeout_ms / 3) && outputPackets_.empty()) {
-					createAttachedNoVersionPacket(ANTOAN_NOP, 0);
-				}
-			}
-		}
-	}
-
-	void readFromSocket() {
-		ActiveLoopWatchdog watchdog(std::chrono::milliseconds(20));
-
-		watchdog.start();
-
-		while (mode_ != ConnectionMode::KILL) {
-			// If the job pool is too busy, do not read more data.
-			if (jobPool_->getJobCount() >= (kMaxBackgroundJobsCount * 9) / 10) { return; }
-
-			uint32_t bytesToRead = inputPacket_.bytesToBeRead();
-			ssize_t ret = ::read(socketFD_, inputPacket_.pointerToBeReadInto(), bytesToRead);
-
-			if (ret == 0) {
-				safs::log_info("connection reset by Master");
-				mode_ = ConnectionMode::KILL;
-				return;
-			}
-
-			if (ret < 0) {
-				if (errno != EAGAIN) {
-					safs::log_error_code(errno, "read from Master error");
-					mode_ = ConnectionMode::KILL;
-				}
-				return;
-			}
-
-			bytesIn_ += ret;
-
-			try {
-				inputPacket_.increaseBytesRead(ret);
-			} catch (InputPacketTooLongException &ex) {
-				safs::log_warn("reading from master: {}", ex.what());
-				mode_ = ConnectionMode::KILL;
-				return;
-			}
-
-			if (ret == bytesToRead && !inputPacket_.hasData()) {
-				continue;  // there might be more data to read in socket's buffer
-			}
-
-			if (!inputPacket_.hasData()) { return; }
-
-			// We have a complete packet in the input buffer, let's process it.
-			gotPacket(inputPacket_.getHeader(), inputPacket_.getData());
-
-			inputPacket_.reset();
-
-			if (watchdog.expired()) { break; }
-		}
-	}
-
-	void writeToSocket() {
-		ActiveLoopWatchdog watchdog(std::chrono::milliseconds(20));
-		ssize_t bytesWritten{-1};
-
-		watchdog.start();
-
-		while (!outputPackets_.empty()) {
-			OutputPacket &pack = outputPackets_.front();
-			bytesWritten = ::write(socketFD_, pack.packet.data() + pack.bytesSent,
-			                       pack.packet.size() - pack.bytesSent);
-
-			if (bytesWritten < 0) {
-				if (errno != EAGAIN) {
-					safs::log_error_code(errno, "write to Master error");
-					mode_ = ConnectionMode::KILL;
-				}
-				return;
-			}
-
-			bytesOut_ += bytesWritten;
-			pack.bytesSent += bytesWritten;
-
-			if (pack.packet.size() != pack.bytesSent) { return; }
-
-			outputPackets_.pop_front();
-
-			if (watchdog.expired()) { break; }
-		}
-	}
-
-	void gotPacket(PacketHeader header, const MessageBuffer &message) try {
-		switch (header.type) {
-		case ANTOAN_NOP:
-			break;
-		case ANTOAN_UNKNOWN_COMMAND:  // for future use
-			break;
-		case ANTOAN_BAD_COMMAND_SIZE:  // for future use
-			break;
-		case SAU_MATOCS_CREATE_CHUNK:
-			createChunk(message);
-			break;
-		case SAU_MATOCS_DELETE_CHUNK:
-			deleteChunk(message);
-			break;
-		case SAU_MATOCS_SET_VERSION:
-			setChunkVersion(message);
-			break;
-		case SAU_MATOCS_DUPLICATE_CHUNK:
-			duplicateChunk(message);
-			break;
-		case SAU_MATOCS_REPLICATE_CHUNK:
-			replicateChunk(message);
-			break;
-		case SAU_MATOCS_TRUNCATE:
-			truncateChunk(message);
-			break;
-		case SAU_MATOCS_DUPTRUNC_CHUNK:
-			duplicateTruncateChunk(message);
-			break;
-		default:
-			safs::log_info("got unknown message (type: {})", header.type);
-			mode_ = ConnectionMode::KILL;
-		}
-	} catch (IncorrectDeserializationException &e) {
-		safs::log_info(
-		    "chunkserver <-> master module: got inconsistent message "
-		    "(type:{}, length:{}), {}",
-		    header.type, uint32_t(message.size()), e.what());
-		mode_ = ConnectionMode::KILL;
-	}
-
-	// Chunk operations
-
-	void createChunk(const std::vector<uint8_t> &data) {
-		uint64_t chunkId;
-		ChunkPartType chunkType = slice_traits::standard::ChunkPartType();
-		uint32_t chunkVersion;
-
-		matocs::createChunk::deserialize(data, chunkId, chunkType, chunkVersion);
-		auto *outputPacket = new OutputPacket;
-		cstoma::createChunk::serialize(outputPacket->packet, chunkId, chunkType, SAUNAFS_STATUS_OK);
-		if (jobPool_) {
-			job_create(*jobPool_, sauJobFinished(this), outputPacket, chunkId, chunkVersion,
-			           chunkType);
-		} else {
-			safs::log_err("MasterConn::createChunk: jobPool is null.");
-			delete outputPacket;
-		}
-	}
-
-	void deleteChunk(const std::vector<uint8_t> &data) {
-		uint64_t chunkId;
-		uint32_t chunkVersion;
-		ChunkPartType chunkType = slice_traits::standard::ChunkPartType();
-
-		matocs::deleteChunk::deserialize(data, chunkId, chunkType, chunkVersion);
-		auto *outputPacket = new OutputPacket;
-		cstoma::deleteChunk::serialize(outputPacket->packet, chunkId, chunkType, 0);
-		if (jobPool_) {
-			job_delete(*jobPool_, sauJobFinished(this), outputPacket, chunkId, chunkVersion,
-			           chunkType);
-		} else {
-			safs::log_err("MasterConn::deleteChunk: jobPool is null.");
-			delete outputPacket;
-		}
-	}
-
-	void setChunkVersion(const std::vector<uint8_t> &data) {
-		uint64_t chunkId;
-		uint32_t chunkVersion;
-		uint32_t newVersion;
-		ChunkPartType chunkType = slice_traits::standard::ChunkPartType();
-
-		matocs::setVersion::deserialize(data, chunkId, chunkType, chunkVersion, newVersion);
-		auto *outputPacket = new OutputPacket;
-		cstoma::setVersion::serialize(outputPacket->packet, chunkId, chunkType, 0);
-		if (jobPool_) {
-			job_version(*jobPool_, sauJobFinished(this), outputPacket, chunkId, chunkVersion,
-			            chunkType, newVersion);
-		} else {
-			safs::log_err("MasterConn::setChunkVersion: jobPool is null.");
-			delete outputPacket;
-		}
-	}
-
-	void duplicateChunk(const std::vector<uint8_t> &data) {
-		uint64_t newChunkId, oldChunkId;
-		uint32_t newChunkVersion, oldChunkVersion;
-		ChunkPartType chunkType = slice_traits::standard::ChunkPartType();
-
-		matocs::duplicateChunk::deserialize(data, newChunkId, newChunkVersion, chunkType,
-		                                    oldChunkId, oldChunkVersion);
-		auto *outputPacket = new OutputPacket;
-		cstoma::duplicateChunk::serialize(outputPacket->packet, newChunkId, chunkType, 0);
-		if (jobPool_) {
-			job_duplicate(*jobPool_, sauJobFinished(this), outputPacket, oldChunkId,
-			              oldChunkVersion, oldChunkVersion, chunkType, newChunkId, newChunkVersion);
-		} else {
-			safs::log_err("MasterConn::duplicateChunk: jobPool is null.");
-			delete outputPacket;
-		}
-	}
-
-	void truncateChunk(const std::vector<uint8_t> &data) {
-		uint64_t chunkId;
-		ChunkPartType chunkType = slice_traits::standard::ChunkPartType();
-		uint32_t version;
-		uint32_t chunkLength;
-		uint32_t newVersion;
-
-		matocs::truncateChunk::deserialize(data, chunkId, chunkType, chunkLength, newVersion,
-		                                   version);
-		auto *outputPacket = new OutputPacket;
-		cstoma::truncate::serialize(outputPacket->packet, chunkId, chunkType, 0);
-		if (jobPool_) {
-			job_truncate(*jobPool_, sauJobFinished(this), outputPacket, chunkId, chunkType, version,
-			             newVersion, chunkLength);
-		} else {
-			safs::log_err("MasterConn::truncateChunk: jobPool is null.");
-			delete outputPacket;
-		}
-	}
-
-	void duplicateTruncateChunk(const std::vector<uint8_t> &data) {
-		uint64_t chunkId, copyChunkId;
-		uint32_t chunkVersion, copyChunkVersion;
-		ChunkPartType chunkType = slice_traits::standard::ChunkPartType();
-		uint32_t newLength;
-
-		matocs::duptruncChunk::deserialize(data, copyChunkId, copyChunkVersion, chunkType, chunkId,
-		                                   chunkVersion, newLength);
-		auto *outputPacket = new OutputPacket;
-		cstoma::duptruncChunk::serialize(outputPacket->packet, copyChunkId, chunkType, 0);
-		if (jobPool_) {
-			job_duptrunc(*jobPool_, sauJobFinished(this), outputPacket, chunkId, chunkVersion,
-			             chunkVersion, chunkType, copyChunkId, copyChunkVersion, newLength);
-		} else {
-			safs::log_err("MasterConn::duplicateTruncateChunk: jobPool is null.");
-			delete outputPacket;
-		}
-	}
-
-	void replicateChunk(const std::vector<uint8_t> &data) {
-		uint64_t chunkId;
-		ChunkPartType chunkType = slice_traits::standard::ChunkPartType();
-		uint32_t chunkVersion;
-		uint32_t sourcesBufferSize;
-		const uint8_t *sourcesBuffer;
-
-		matocs::replicateChunk::deserializePartial(data, chunkId, chunkVersion, chunkType,
-		                                           sourcesBuffer);
-		sourcesBufferSize = data.size() - (sourcesBuffer - data.data());
-
-		auto *outputPacket = new OutputPacket;
-		cstoma::replicateChunk::serialize(outputPacket->packet, chunkId, chunkType,
-		                                  SAUNAFS_STATUS_OK, chunkVersion);
-		safs::log_debug("cs.matocs.replicate {}", chunkId);
-
-		if (hddScansInProgress()) {
-			// Disk scan in progress - replication is not possible
-			sauJobFinished(SAUNAFS_ERROR_WAITING, outputPacket);
-		} else {
-			if (jobPool_) {
-				job_replicate(*jobPool_, sauJobFinished(this), outputPacket, chunkId, chunkVersion,
-				              chunkType, sourcesBufferSize, sourcesBuffer);
-			} else {
-				safs::log_err("MasterConn::replicateChunk: jobPool is null.");
-				delete outputPacket;
-			}
-		}
-	}
-
-	// Callbacks
-
-	static std::function<void(uint8_t status, void *packet)> sauJobFinished(
-	    MasterConn *masterConn) {
-		return [masterConn](uint8_t status, void *packet) {
-			masterConn->sauJobFinished(status, packet);
-		};
-	}
-
-	void sauJobFinished(uint8_t status, void *packet) {
-		auto *outputPacket = static_cast<OutputPacket *>(packet);
-
-		if (mode_ == ConnectionMode::CONNECTED) {
-			cstoma::overwriteStatusField(outputPacket->packet, status);
-			attachPacket(packet);
-		} else {
-			deletePacket(packet);
-		}
-	}
-
-	// Termination
-
-	void releaseResources() {
-		if (mode_ != ConnectionMode::FREE && mode_ != ConnectionMode::CONNECTING) {
-			tcpclose(socketFD_);
-			inputPacket_.reset();
-		}
-	}
-
-	void resetPackets() {
-		inputPacket_.reset();
-		outputPackets_.clear();
-	}
-
-	// Getters and setters
-
-	ConnectionMode mode() const { return mode_; }
-
-	void setMode(ConnectionMode newMode) { mode_ = newMode; }
-
-	const NetworkAddress &address() const { return address_; }
-
-	void setMasterAddress(uint32_t ip_, uint16_t port_) {
-		address_.ip = ip_;
-		address_.port = port_;
-	}
-
-	const NetworkAddress &bindHostAddress() const { return bindHostAddress_; }
-
-	void setBindHostAddress(uint32_t ip_, uint16_t port_) {
-		bindHostAddress_.ip = ip_;
-		bindHostAddress_.port = port_;
-	}
-
-	int socketFD() const { return socketFD_; }
-
-	bool isMasterAddressValid() const { return isMasterAddressValid_; }
-
-	void setMasterAddressValid(bool valid) { isMasterAddressValid_ = valid; }
-
-	uint64_t bytesIn() const { return bytesIn_; }
-	uint64_t bytesOut() const { return bytesOut_; }
-
-	void resetStats() {
-		bytesIn_ = 0;
-		bytesOut_ = 0;
-	}
-
-private:
-	std::string masterHostStr_;                  ///< Hostname of the master server.
-	std::string masterPortStr_;                  ///< Port of the master server.
-	std::shared_ptr<JobPool> jobPool_;           ///< Shared reference to the JobPool.
-
-	ConnectionMode mode_{ConnectionMode::FREE};  ///< Current mode of the connection to this master.
-	int socketFD_{-1};                           ///< Socket file descriptor for this connection.
-	int32_t pDescPos_{-1};                       ///< Position in the pollfd array.
-	Timer lastRead_;                             ///< Time since the last read operation.
-	Timer lastWrite_;                            ///< Time since the last write operation.
-	InputPacket inputPacket_{kMaxPacketSize};    ///< Input buffer for reading data from the socket.
-	std::list<OutputPacket> outputPackets_;      ///< Output packets to be sent to the master.
-
-
-	NetworkAddress address_;            ///< Address of this master server (IP and port).
-	NetworkAddress bindHostAddress_;    ///< Address to bind the socket to (IP and port).
-	bool isMasterAddressValid_{false};  ///< Tells if the master address is valid.
-
-	// Statistics
-	uint64_t bytesIn_ = 0;   ///< Number of bytes read from the master.
-	uint64_t bytesOut_ = 0;  ///< Number of bytes sent to the master.
-};
 
 //  From config
 static std::string gMasterHost;
 static std::string gMasterPort;
+static bool gEnableLoadFactor;
 
 static const uint64_t kSendStatusDelay = 5;
